feat: Absence controller fully operationnal + Exception management done

+++ api-back/src/main/java/com/diginamic/apiback/controllers/AbsenceController.java
@@ -3,6 +3,8 @@ package com.diginamic.apiback.controllers;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.lang.NonNull;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,11 +14,15 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.ArrayList;
 import java.util.List;
 
+import com.diginamic.apiback.dto.AbsenceDTO;
 import com.diginamic.apiback.models.Absence;
 import com.diginamic.apiback.services.AbsenceService;
 
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
 
 @RestController
@@ -24,31 +30,53 @@ import jakarta.validation.Valid;
 public class AbsenceController {
     @Autowired
     AbsenceService absenceService;
-    
-        @GetMapping()
-    public List<Absence> findAll() {
-        return absenceService.findAll();
+
+    @GetMapping()
+    public List<AbsenceDTO> findAll() {
+        List<AbsenceDTO> absenceDtoList = new ArrayList<>();
+        List<Absence> absenceList = absenceService.findAll();
+        for (Absence absence : absenceList) {
+            absenceDtoList.add(absence.toDto());
+        }
+        return absenceDtoList;
     }
 
     @GetMapping("/{id}")
-    public Optional<Absence> findById(@NonNull @PathVariable("id") Long id) {
-        return absenceService.findById(id);
+    public AbsenceDTO findById(@NonNull @PathVariable("id") Long id) {
+        Optional<Absence> absence = absenceService.findById(id);
+        if (absence.isPresent()) {
+            return absence.get().toDto();
+        } else {
+            throw new EntityNotFoundException("No absence entity with corresponding id found in db");
+        }
     }
 
     // @CrossOrigin(origins = "http://localhost:4200")
     @PostMapping("/create")
-    public Absence createAbsence(@RequestBody @Valid Absence absence) {
-        return absenceService.createAbsence(absence);
+    public AbsenceDTO createAbsence(@RequestBody @Valid Absence absence) {
+        Absence abs = absenceService.createAbsence(absence);
+        return abs.toDto();
     }
 
-    @PutMapping("/{id}")
-    public Absence updateUser(@NonNull @RequestBody @Valid Absence absence, @NonNull @PathVariable("id")Long id){
-        return absenceService.updateAbsence(absence, id);
+    @PutMapping("/update/{id}")
+    public AbsenceDTO updateUser(@NonNull @RequestBody @Valid Absence absence, @NonNull @PathVariable("id") Long id) {
+        Optional<Absence> abs = absenceService.findById(id);
+        if (abs.isPresent()) {
+            return absenceService.updateAbsence(absence, id).toDto();
+        } else {
+            throw new EntityNotFoundException("No absence entity with corresponding id found in db");
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public Absence deleteUser(@NonNull @PathVariable("id")Long id){
-       return absenceService.deleteAbsence(id);
+    public AbsenceDTO deleteUser(@NonNull @PathVariable("id") Long id) {
+        Optional<Absence> abs = absenceService.findById(id);
+        if (abs.isPresent()) {
+            absenceService.deleteAbsence(id);
+            return abs.get().toDto();
+        } else {
+            throw new EntityNotFoundException("No absence entity with corresponding id found in db");
+        }
     }
 
 }
